fix: Moving `AuthProviderCmd` tests that require AWS access to aws build flag gatted files
